+++ platform_gp/types.go
@@ -1,1 +1,4 @@
 package platform_gp
+
+type GoldenPath struct {
+}
