--- rules/dry_run.go
@@ -1,3 +1,0 @@
-package rules
-
-// code regarding performing a dry-run against the APIs based on the defined request contract
